Fix a bad merge.

+++ core/src/sigverifier/bls_sigverifier.rs
@@ -144,9 +144,13 @@ impl BLSSigVerifier {
         }
         datapoint_info!(
             "bls_sig_verifier_stats",
-            ("sent", self.stats.sent, u64),
-            ("sent_failed", self.stats.sent_failed, u64),
-            ("verified_votes_sent", self.stats.verified_votes_sent, u64),
+            ("sent", self.stats.sent as i64, i64),
+            ("sent_failed", self.stats.sent_failed as i64, i64),
+            (
+                "verified_votes_sent",
+                self.stats.verified_votes_sent as i64,
+                i64
+            ),
             (
                 "verified_votes_sent_failed",
                 self.stats.verified_votes_sent_failed,
